# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "My Fucking Swaggah Jagahh.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/deleteSavedAccount": {
            "delete": {
                "description": "DeleteSavedAccount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SavedAccount"
                ],
                "summary": "DeleteSavedAccount",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeletePreferredAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/getSavedAccountbyCid": {
            "post": {
                "description": "GetSavedAccountbyCid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SavedAccount"
                ],
                "summary": "GetSavedAccountbyCid",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetSaveAcc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PreferredFT"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/getSavedAccountbyCidAndtargetCid": {
            "post": {
                "description": "GetSavedAccount by Cid and target Cid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SavedAccount"
                ],
                "summary": "GetSavedAccount by Cid and target Cid",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetSaveAccAndTCid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PreferredFT"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/updateSavedAccount": {
            "post": {
                "description": "Fields are optional except cid and target cid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SavedAccount"
                ],
                "summary": "UpdateSavedAccount",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PreferredFT"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PreferredFT"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DeletePreferredAccount": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "target_cid": {
                    "type": "string"
                }
            }
        },
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "retcode": {
                    "type": "integer"
                }
            }
        },
...
        "model.GetSaveAcc": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                }
            }
        },
        "model.GetSaveAccAndTCid": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "target_cid": {
                    "type": "string"
                }
            }
        },
...
        "model.PreferredFT": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "cid": {
                    "type": "string"
                },
                "date_added": {
                    "type": "string"
                },
                "insti": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "target_cid": {
                    "type": "string"
                }
            }
        },
...
    }
}